refactor: adapt changes to the lates `xo`
