+++ .travis.yml
@@ -4,12 +4,7 @@ notifications:
 node_js:
   - '8'
   - '6'
-script:
-  - npm run lint
-  - node node_modules/.bin/prettier -l "**/*.{js,json,md}"
-  - npm test
-after_success:
-  - if [[ `npm -v` > 4* ]]; then npx codecov; fi
+after_success: if [[ `npm -v` > 4* ]]; then npx codecov; fi
 branches:
   only:
     - master
